refactor: api 파일 위치가 실제 route와 동일하게 변경
